+++ src/ParamConverter/FloatParamConverter.php
@@ -1,24 +1,24 @@
 <?php
 
-namespace ParamConverter;
+namespace ParamConverter\ParamConverter;
 
 use Cake\Http\Exception\BadRequestException;
 
 /**
- * Class EntityParamConverter
+ * Class FloatParamConverter
  *
- * Param Converter for converting integer-like request strings to int Controller parameters
+ * Param Converter for converting float like request strings to float Controller parameters
  *
  * @package ParamConverter
  */
-class IntegerParamConverter implements ParamConverterInterface
+class FloatParamConverter implements ParamConverterInterface
 {
     /**
      * @inheritDoc
      */
     public function supports(string $class): bool
     {
-        return $class === 'int';
+        return $class === 'float';
     }
 
     /**
@@ -26,8 +26,8 @@ class IntegerParamConverter implements ParamConverterInterface
      */
     public function convertTo(string $value, string $class)
     {
-        if (ctype_digit($value)) {
-            return (int)$value;
+        if (is_numeric($value)) {
+            return (float)$value;
         }
         throw new BadRequestException();
     }
