+++ src/ParamConverter/IntegerParamConverter.php
@@ -1,24 +1,24 @@
 <?php
 
-namespace ParamConverter;
+namespace ParamConverter\ParamConverter;
 
 use Cake\Http\Exception\BadRequestException;
 
 /**
- * Class FloatParamConverter
+ * Class EntityParamConverter
  *
- * Param Converter for converting float like request strings to float Controller parameters
+ * Param Converter for converting integer-like request strings to int Controller parameters
  *
  * @package ParamConverter
  */
-class FloatParamConverter implements ParamConverterInterface
+class IntegerParamConverter implements ParamConverterInterface
 {
     /**
      * @inheritDoc
      */
     public function supports(string $class): bool
     {
-        return $class === 'float';
+        return $class === 'int';
     }
 
     /**
@@ -26,8 +26,8 @@ class FloatParamConverter implements ParamConverterInterface
      */
     public function convertTo(string $value, string $class)
     {
-        if (is_numeric($value)) {
-            return (float)$value;
+        if (ctype_digit($value)) {
+            return (int)$value;
         }
         throw new BadRequestException();
     }
